Software-Engineering-Team-main/software_engineering_team/requirements.txt
# Core dependencies
crewai>=0.28.0
litellm>=1.30.0
python-dotenv>=1.0.0

# UI and file processing
gradio>=4.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Additional utilities
pathlib2>=2.3.7
zipfile36>=0.1.3
requests>=2.31.0

# Optional: For better debugging
rich>=13.0.0
colorama>=0.4.6 